Added some files and pimped backup

+++ bin/backup_pg.sh
@@ -1,19 +1,45 @@
 #!/usr/bin/env bash
 
-source ~/.prv/env
+# Usage: ./pg_backup.sh -h <host> -u <user> -p <password> -d <backup_dir>
+# Example: ./pg_backup.sh -h localhost -u postgres -p password -d /mnt/storage/backups/postgres/
+OPTSRING=":h:u:p:d:"
+while getopts $OPTSRING opt; do
+    case $opt in
+    h)
+        host=$OPTARG
+        ;;
+    u)
+        user=$OPTARG
+        ;;
+    p)
+        password=$OPTARG
+        ;;
+    d)
+        backup_dir=$OPTARG
+        ;;
+    \?)
+        echo "Invalid option: -$OPTARG" >&2
+        ;;
+    :)
+        echo "Option -$OPTARG requires an argument." >&2
+        ;;
+    esac
+done
 
-backup_dir="/mnt/storage/backups/postgres/kubes/"
-nightly_dir="/mnt/storage/backups/postgres/latest/"
-backup_date=`date +%d-%m-%Y_%H:%M`
+if [ -z "$host" ] || [ -z "$user" ] || [ -z "$password" ] || [ -z "$backup_dir" ]; then
+    echo 'Missing required arguments' >&2
+    exit 1
+fi
+
+backup_date=$(date +%d-%m-%Y_%H:%M)
 number_of_days=15
-host="daphne.l.fergl.ie"
-user=postgres
 
-databases=`PGPASSWORD=$CLUSTER_MASTER_PGPWD psql -h "$host" -U $user -l -t | cut -d'|' -f1 | sed -e 's/ //g' -e '/^$/d'`
-for i in $databases; do  if [ "$i" != "template0" ] && [ "$i" != "template1" ] && [ "$i" != "template_postgis" ]; then    
-    echo Dumping $i to $backup_dir$i\_$backup_date.sql    
-    PGPASSWORD=$CLUSTER_MASTER_PGPWD pg_dump -h $host -U $user $i > $backup_dir$i\_$backup_date.sql
-    bzip2 $backup_dir$i\_$backup_date.sql
-  fi
+databases=$(PGPASSWORD=$password psql -h "$host" -U $user -l -t | cut -d'|' -f1 | sed -e 's/ //g' -e '/^$/d')
+for i in $databases; do
+    if [ "$i" != "template0" ] && [ "$i" != "template1" ] && [ "$i" != "template_postgis" ]; then
+        echo Dumping $i to $backup_dir$i\_$backup_date.sql
+        PGPASSWORD=$password pg_dump -h $host -U $user $i >$backup_dir$i\_$backup_date.sql
+        bzip2 $backup_dir$i\_$backup_date.sql
+    fi
 done
 find $backup_dir -type f -prune -mtime +$number_of_days -exec rm -f {} \;
